sip: fix Authorization header with mixed quoted parameters
Fix to correctly read unquoted parameter values in Authorization lines.
Code in master branch will look for double quotes as the end of a
parameter value even in case the value is not enclosed in double quotes.
This fix checks if the first character in a value is a double quote in
which case it will scan until a non-escaped finishing double quote is
found to determine the end of the value string. If no double quote is
found as the start of the parameter value it will scan for a comma as
the end of the value string.

Change-Id: I5322caa146d3213bc94ad24528c2fa23cdea10f2
Bug: 13035
Reviewed-on: https://code.wireshark.org/review/18344
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Roland Knall <rknall@gmail.com>