WelcomeUI: Enhance Open and Capture clickable labels
Open and Capture are now clickable labels, but they are sensitive
throughout the whole width of the central column. Without background
highlighting, as in the file list, this may be non-intuative.
To more closely resemble a button, limit the sensitivity to the text
of the label only, by limiting the label size.

Change-Id: I0b703af02abc71a48dfbcee81f336bba955b4a70
Reviewed-on: https://code.wireshark.org/review/18349
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Roland Knall <rknall@gmail.com>