Add .gitlab-ci.yml.
As for .travis.yml, this file triggers the CI module of GITLAB.
On the public instance of gitlab (gitlab.com) the runners are active
by default, then any mirror of wireshark on gitlab.com will result
in an automated build.

The current version of .gitlab-ci.yml builds with gcc 4,5,6 and with
default config, without extcap and without pcap.

Change-Id: I0800045e7f1b4a137dbd69a185d3b09a89398948
Reviewed-on: https://code.wireshark.org/review/18936
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Reviewed-by: Dario Lombardo <lomato@gmail.com>