tpncp: load hf array in the handoff function
tpncp reads hf entries from a database. There's a boolean preference
defining if the database is actually read. We can't access this
preference until the preferences are initialized. The code as it is now
wil always default to false and never read the database.

Check the preference in the handoff function and initialize the hf
entries if they are not yet initialized.

Change-Id: I545a49a946b78e1a0bc23e9803eb671d8765bba4
Reviewed-on: https://code.wireshark.org/review/18386
Reviewed-by: Martin Kaiser <wireshark@kaiser.cx>
Petri-Dish: Martin Kaiser <wireshark@kaiser.cx>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>