POP: replace wmem_strndup by tvb_memcpy
gdd3b977 introduced the use of wmem_strndup to ensure that the string is
NULL terminated.
But it also stops as soon as it finds a NULL character, which can happen
with malformed packets. So let's copy all the bytes with tvb_memcpy and
add manually a NULL character at the end of the buffer for safeness.

Bug: 13074
Change-Id: I9ae65d739a92bc6167e4acfc275e1909f3890815
Reviewed-on: https://code.wireshark.org/review/18582
Petri-Dish: Pascal Quantin <pascal.quantin@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Pascal Quantin <pascal.quantin@gmail.com>