autotools: propagate git version info to distribution tarball
Store the git describe output in the distribution tarball (make dist)
and ensure that this version is always used for builds from this
tarball. This will prevent the useless "Git Rev Unknown from unknown"
output in tar-based builds.

It will also prevent git from being invoked in tarball builds.

Remove the git branch name since the commit ID (and git tag) in the git
describe output is sufficient to identify the source tree. (In SVN, a
revision ID had to be paired with a branch name to identify the source
tree, in git this is no longer the case.)

Change-Id: Iffe142b6efd81e857802eb604d6310cfd301d207
Reviewed-on: https://code.wireshark.org/review/18415
Reviewed-by: João Valverde <j@v6e.pt>
Reviewed-by: Peter Wu <peter@lekensteyn.nl>
Petri-Dish: Peter Wu <peter@lekensteyn.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Dario Lombardo <lomato@gmail.com>