Fix some errors.
Th link throughput and latency objects are just 32-bit integers; process
them as such and, as they're bytes-per-second and microsecond values,
respectively, show them as decimal, rather than as hex, as most of us
have 10, not 16, fingers.

The ETX metric object is a 16-bit integer; process it as such, and show
it in decimal as well.

Attach an expert info representing an error to the item in error.

Change-Id: I57dfac0127d263724cc6940a0457d63e17616b64
Reviewed-on: https://code.wireshark.org/review/18617
Reviewed-by: Guy Harris <guy@alum.mit.edu>