debian-setup: allow libssh-dev to be installed for trusty
Ubuntu 14.04 Trusty has libssh-dev 0.6.1 (using OpenSSL) which should be
usable.

The debian/control file is not updated due to OpenSSL licensing
concerns, if you need to build on trusty, you need dpkg-buildpackage -d
(--no-check-builddeps).

Remove distro-detection in favor of feature detection. Note that the
minimum Ubuntu versions could be lower, but unsupported versions are not
listed on packages.ubuntu.com so I did not bother.

Change-Id: Ia04d960f3d4957918c925bf9cd351537e233fd3e
Reviewed-on: https://code.wireshark.org/review/18725
Reviewed-by: Dario Lombardo <lomato@gmail.com>
Reviewed-by: Balint Reczey <balint@balintreczey.hu>