Qt: populate Bookmarks menu in capture options dialog
The bookmarks menu is populated for the main window when the
appInitialized signal is emitted, this ensures that any preferences are
applied.

For the Capture Options dialog however, the appInitialized signal is
never triggered since the GUI has already started. Just invoke it from
the constructor such the menu is populated at least once.

Bug: 12986
Change-Id: I04457fd25e52a25714ac471131b90e3a128515d8
Reviewed-on: https://code.wireshark.org/review/18734
Reviewed-by: Jaap Keuter <jaap.keuter@xs4all.nl>
Petri-Dish: Jaap Keuter <jaap.keuter@xs4all.nl>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>