Don't assume we have a protocol tree.
In the association_sanity_check_t structure:

Have separate flags to indicate whether we *have* an IE specifying an
FT AKM suite and whether we *have* an IE specifying a non-FT AKM suite.
Use those to decide whether to add an expert info, so we add it
regardless of whether we are building a protocol tree or not.

Set the protocol tree item for the first FT AKM suite and the first
non-FT AKM suite only if we have a protocol tree.

Bug: 13149
Change-Id: I16a3194017b34d8d2c9128364c2761d1c202969d
Reviewed-on: https://code.wireshark.org/review/18863
Reviewed-by: Guy Harris <guy@alum.mit.edu>