The version lists are sequences of 1-byte integers, not strings.
Show each version in the list independently as an item.

Perhaps the Set Protocol response version lists seen have only one
version, but the presence of a version-0 terminator suggests that it
could contain multiple versions, so dissect it as such.

For FT_STRINGZ values, let proto_tree_add_item() determine the length -
pass a length of -1.  If we need the length, use
proto_tree_add_item_ret_length().

Change-Id: I5954ccac34f9e462c6d43e9a213974cf818f4d0d
Reviewed-on: https://code.wireshark.org/review/19134
Reviewed-by: Guy Harris <guy@alum.mit.edu>