VLAN: Support IEEE 802.1Q-2005 and -2011 changes
Supports the recommended meanings of the priority field as
changed by 802.1Q-2005, and the change from the CFI to DEI
in 802.1Q-2011. A preference is added to use the older
(non-compatible) spec version. Note that 802.1Q-2011 is
consistent with 802.1ad and ah, which got rolled up in 802.1Q,
but ends up removing support for bridging Token Ring and FDDI
over Ethernet.

Bug: 13294
Change-Id: Ieeadb0f6dda2758750f9e6649f1390609d78c50e
Reviewed-on: https://code.wireshark.org/review/19548
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>