802.11(ad): Fix SSW Feedback dissection when Sector Sweep Direction = Responder
For the Sector Sweep Frame in which the Sector Sweep Direction = Responder, the format of the Sector Sweep Feedback Field should be the one in the standard (Figure 8-431d—SSW Feedback field format when not transmitted as part of an ISS) i.e. similar to the one in Sector Sweep Feedback Frame and Acknowledgement.

Issue reported by Hany ASSASA

bug: 13244
Change-Id: Ic8c6d83fc32d017fb73116a54759608498f99452
Reviewed-on: https://code.wireshark.org/review/19590
Reviewed-by: Anders Broman <a.broman58@gmail.com>