mswsp: Numerous bugfixes
1. Fix display of CTableColumn.StatusUsed & CTableColumn.LengthUsed
Both StatusUsed and LenghtUsed members were displayed as 2 byte length
values instead of single byte values.

2. Fix minor spelling mistake

3. display correct address(s) for array of address values

4. Use correct minimum CTableColumn size

This is followup patch for patch already provided for bug 13299.
The previous minimum value neglated to take into account the
CFullPropSpec structure (which is on the wire) but is represtented
as a string in the internal CTableColumn structure. Note: this doesn't
make any difference to the effectiveness of the previous patch but should
avoid confusion if someone later actually investigates the correct minimum
size.

5. Fix CPMCREATEQUERY message parsing

Following a query and subsequent clarification from dochelp@microsoft.com
the MS-WSP specification document was found to be incorrect.

CPMCreateQueryIn
      |
      ---> CSortSet

should instead eventually point to CSortSet via the following msg structure

CPMCreateQueryIn
      |
      ---> CInGroupSortAggregSets
           |
            ---> CInGroupSortAggregSet
                 |
                 ---> CSortSet

Change-Id: I7da6c7db817113f5542f25f078626030ad30d737
Signed-off-by: Noel Power <noel.power@suse.com>
Reviewed-on: https://code.wireshark.org/review/20031
Reviewed-by: Michael Mann <mmann78@netscape.net>