Logcat-text: use GRegex optimizations
"G_REGEX_OPTIMIZE - Optimize the regular expression. If the pattern
will be used many times, then it may be worth the effort to optimize
it to improve the speed of matches." - Glib documentation.

It is possible to capture a lot of Logcat logs or these log may
flooding us. Optimizations are welcome.

Change-Id: If753e795efe30b014a5fad11c8ebbcd4da3824a6
Reviewed-on: https://code.wireshark.org/review/20357
Petri-Dish: Michael Mann <mmann78@netscape.net>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Michael Mann <mmann78@netscape.net>