Qt: Add check for valid profile name
Add check for a valid profile name in the Profile dialog and disable
the Ok button and mark the invalid profile if error.

The GTK dialog does not close on errors but the Qt dialog does, so
ensure the user is informed beforehand.

Change-Id: If31ffe75183066cf49e2b5a082112bfe62d6c06e
Reviewed-on: https://code.wireshark.org/review/19995
Petri-Dish: Stig Bjørlykke <stig@bjorlykke.org>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Stig Bjørlykke <stig@bjorlykke.org>