Reduce number of fwrite when writing pcap-ng EPB.
According to callgrind with the patch dumping 3000 pakets the number of
calls to write_to_file is reduced from 11541 to 9000 reducing the number
of lr from 4 681 518 to 4 314 101.

If the buffer holding the packet was guaranteed to be padded to 32 bit
boundary the code could be simplified and if there was space "in front"
for the packet header it would be even better.

Change-Id: Ie991c05fa9d831ee4d703bd47b8123f2b1f83277
Reviewed-on: https://code.wireshark.org/review/20256
Petri-Dish: Anders Broman <a.broman58@gmail.com>
Tested-by: Petri Dish Buildbot <buildbot-no-reply@wireshark.org>
Reviewed-by: Anders Broman <a.broman58@gmail.com>